+++ default.env
@@ -85,3 +85,16 @@ export E2E_PAC_GITHUB_APP_ID=''
 # Note: How to get private key https://docs.github.com/en/apps/creating-github-apps/authenticating-with-a-github-app/managing-private-keys-for-github-apps
 # Required: only for running tests that are using PaC (see README.md for more details)
 export E2E_PAC_GITHUB_APP_PRIVATE_KEY=''
+
+# Only for upgrade tests
+# Branch with changes for upgrade (eg. new image tag...)
+# Example: quality-dashboard
+# Required: yes (for upgrade tests)
+export UPGRADE_BRANCH=quality-dashboard
+
+# Only for upgrade tests
+# Branch with changes for upgrade (eg. new image tag...)
+# Example: redhat-appstudio-qe
+# Required: no
+# Default value(if not specified): redhat-appstudio
+export UPGRADE_FORK_ORGANIZATION=redhat-appstudio-qe
